caching no longer used (code still there but commented out), made performance very bad for all output media progress bar for writing is now in % mode, seems to work cancelling writing really does work now
git-svn-id: svn://littlesvr.ca/prog/isomaster@244 ea4051ab-32fe-0310-aa8c-8a8c8f7b8cea

+++ src/browser.c
@@ -149,7 +149,7 @@ void createDirCbk(GtkButton *button, gpointer onFs)
     gtk_widget_destroy(dialog);
 }
 
-void formatSize(unsigned long long  sizeInt, char* sizeStr, int sizeStrLen)
+void formatSize(off_t sizeInt, char* sizeStr, int sizeStrLen)
 {
     if(sizeInt > 1073741824)
     /* print gibibytes */
